refactor(cms-base-nestjs-module): remove signature service
