+++ packages/cms-base-nestjs-module/src/index.ts
@@ -4,7 +4,6 @@ export * from './cms-base.module';
 // Services
 export * from './services/article-base.service';
 export * from './services/category-base.service';
-export * from './services/article-signature.service';
 
 // DataLoaders
 export * from './data-loaders/category.dataloader';
